添加MQ安全链路

+++ NBlogBackEnd/src/main/java/com/nblog/consumer/ArticleConsumer.java
@@ -7,8 +7,11 @@ import com.nblog.dto.ArticleDTO;
 import com.nblog.entity.Article;
 import com.nblog.enums.SearchStrategyEnum;
 import com.nblog.service.ESService;
+import com.nblog.service.RedisService;
 import com.nblog.utils.CommonUtils;
+import com.rabbitmq.client.Channel;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.amqp.core.Message;
 import org.springframework.amqp.rabbit.annotation.RabbitHandler;
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,6 +26,8 @@ import java.util.Date;
 
 import static com.nblog.constant.CommonConst.MDS_PATH;
 import static com.nblog.constant.MQPrefixConst.ES_QUEUE;
+import static com.nblog.constant.RedisPrefixConst.ARTICLE_MESSAGE;
+import static com.nblog.constant.RedisPrefixConst.CONSUMER_ARTICLE_MESSAGE;
 
 /**
  * @author liulin
@@ -36,6 +41,7 @@ public class ArticleConsumer {
     private final UserDao userDao;
     private final ArticleDao articleDao;
     private final ESService esService;
+    private final RedisService redisService;
 
     /**
      * 构造器方式注入
@@ -44,10 +50,11 @@ public class ArticleConsumer {
      * @param esService
      */
     @Autowired
-    public ArticleConsumer(UserDao userDao,ArticleDao articleDao,ESService esService){
+    public ArticleConsumer(UserDao userDao,ArticleDao articleDao,ESService esService,RedisService redisService){
         this.userDao = userDao;
         this.articleDao=articleDao;
         this.esService=esService;
+        this.redisService =redisService;
     }
 
     // 用 ThreadLocal 保证 SimpleDateFormat 线程安全
@@ -56,23 +63,49 @@ public class ArticleConsumer {
 
     /**
      * rabbitmq article 消费者
-     * @param art
      */
     @RabbitHandler
-    public void process(String art){
-        Article article = JSON.parseObject(art,Article.class);
+    public void process(String msg, Channel channel, Message message){
+        // 获取消息 UUID
+        String uuid = (String) message.getMessageProperties().getHeaders().get("spring_returned_message_correlation");
 
-        // 根据用户 id 来区分是添加文档还是删除文档
-        if(article.getUserId()!=-1){
-            // 添加文档
-            addArticle(article);
-        }else {
-            // 删除文档
-            esService.deleteDocument(SearchStrategyEnum.NBLOG.getDesc(),article.getId()+"");
+        // 幂等性，防止重复消费
+        if(redisService.sIsMember(CONSUMER_ARTICLE_MESSAGE,uuid)){
+            // 待消费的文章
+            Article article = JSON.parseObject(msg,Article.class);
+
+            // 根据用户 id 来区分是添加文档还是删除文档
+            if(article.getUserId()!=-1){
+                // 添加文档
+                addArticle(article);
+            }else {
+                // 删除文档
+                esService.deleteDocument(SearchStrategyEnum.NBLOG.getDesc(),article.getId()+"");
+            }
+
+            // 从 Redis 中删除待消费
+            redisService.sRemove(CONSUMER_ARTICLE_MESSAGE,uuid);
+            // 从 Redis 中删除该条消息
+            redisService.hDel(ARTICLE_MESSAGE,uuid);
         }
 
-        // 文章已消费，打印日志
-        log.info("消费文章：" + article.toString());
+        try {
+            // 手动执行 ack
+            channel.basicAck(message.getMessageProperties().getDeliveryTag(), false);
+            // 文章已消费，打印日志
+            log.info("【消费者】： 成功消费消息 " + uuid);
+
+        } catch (IOException e) {
+            //消费者处理出了问题，需要告诉队列信息消费失败
+            try {
+                channel.basicNack(message.getMessageProperties().getDeliveryTag(),
+                        false, true);
+            } catch (IOException ex) {
+                log.error(ex.getMessage());
+            }
+
+            System.err.println("【消费者】： 消费消息 "+ uuid +"失败");
+        }
     }
 
     /**
@@ -82,7 +115,6 @@ public class ArticleConsumer {
      */
     private void addArticle(Article article){
         String content = article.getArticleContent();
-
         // 写入磁盘
         String base= MDS_PATH + article.getUserId() + File.separator;
         String fileName=article.getArticleTitle()+"_"+article.getId();
